services/auth/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
sqlmodel==0.0.8
asyncpg==0.28.0
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
redis==5.0.1
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
argon2-cffi==23.1.0
email-validator==2.1.0
tenacity==8.2.3
structlog==23.2.0 